backend/pyproject.toml
[tool.poetry]
name = "freshcart"
version = "0.1.0"
description = "Shop like you never did before"
authors = ["SCADEMY <dev@scademy.com>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "3.11.5"
Flask = "3.0.0"
alembic = "1.12.1"
psycopg2 = "2.9.9"
openai = "1.3.4"
Flask-SocketIO = "5.3.6"
Flask-SQLAlchemy = "3.1.1"
pgvector = "0.2.3"
Flask-Cors = "4.0.0"
Flask-JWT-Extended = "4.5.3"
click = "8.1.7"

[tool.poetry.dev-dependencies]
black = "23.11.0"
isort = "5.12.0"
autoflake = "2.2.1"
nose2 = "0.14.0"
pytruth = "1.1.0"

[tool.poetry.scripts]
embed = 'freshcart.scripts.embed:main'

[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
